requirements.txt
python-telegram-bot[job-queue]
supabase
anthropic
python-dotenv
